break out slack.go into more files so that it is less confusing
